+++ controller_interface/src/controller_interface.cpp
@@ -22,12 +22,14 @@ ControllerInterface::ControllerInterface() : ControllerInterfaceBase() {}
 
 bool ControllerInterface::is_chainable() const { return false; }
 
-std::vector<hardware_interface::StateInterface> ControllerInterface::export_state_interfaces()
+std::vector<hardware_interface::StateInterface::ConstSharedPtr>
+ControllerInterface::export_state_interfaces()
 {
   return {};
 }
 
-std::vector<hardware_interface::CommandInterface> ControllerInterface::export_reference_interfaces()
+std::vector<hardware_interface::CommandInterface::SharedPtr>
+ControllerInterface::export_reference_interfaces()
 {
   return {};
 }