+++ hardware_interface/include/hardware_interface/sensor_interface.hpp
@@ -15,9 +15,14 @@
 #ifndef HARDWARE_INTERFACE__SENSOR_INTERFACE_HPP_
 #define HARDWARE_INTERFACE__SENSOR_INTERFACE_HPP_
 
+#include <limits>
+#include <memory>
 #include <string>
+#include <unordered_map>
+#include <utility>
 #include <vector>
 
+#include "hardware_interface/component_parser.hpp"
 #include "hardware_interface/handle.hpp"
 #include "hardware_interface/hardware_info.hpp"
 #include "hardware_interface/types/hardware_interface_return_values.hpp"
@@ -115,21 +120,103 @@ public:
    * \returns CallbackReturn::SUCCESS if required data are provided and can be parsed.
    * \returns CallbackReturn::ERROR if any error happens or data are missing.
    */
-  virtual CallbackReturn on_init(const HardwareInfo & /*hardware_info*/)
+  virtual CallbackReturn on_init(const HardwareInfo & hardware_info)
   {
+    info_ = hardware_info;
+    import_state_interface_descriptions(info_);
     return CallbackReturn::SUCCESS;
   };
 
+  /**
+   * Import the InterfaceDescription for the StateInterfaces from the HardwareInfo.
+   * Separate them into the possible types: Sensor and store them.
+   */
+  virtual void import_state_interface_descriptions(const HardwareInfo & hardware_info)
+  {
+    auto sensor_state_interface_descriptions =
+      parse_state_interface_descriptions(hardware_info.sensors);
+    for (const auto & description : sensor_state_interface_descriptions)
+    {
+      sensor_state_interfaces_.insert(std::make_pair(description.get_name(), description));
+    }
+  }
+
   /// Exports all state interfaces for this hardware interface.
   /**
-   * The state interfaces have to be created and transferred according
-   * to the hardware info passed in for the configuration.
+   * Old way of exporting the StateInterfaces. If a empty vector is returned then
+   * the on_export_state_interfaces() method is called. If a vector with StateInterfaces is returned
+   * then the exporting of the StateInterfaces is only done with this function and the ownership is
+   * transferred to the resource manager. The set_command(...), get_command(...), ..., can then not
+   * be used.
    *
    * Note the ownership over the state interfaces is transferred to the caller.
    *
    * \return vector of state interfaces
    */
-  virtual std::vector<StateInterface> export_state_interfaces() = 0;
+  [[deprecated(
+    "Replaced by vector<StateInterface::ConstSharedPtr> on_export_state_interfaces() method. "
+    "Exporting is handled "
+    "by the Framework.")]] virtual std::vector<StateInterface>
+  export_state_interfaces()
+  {
+    // return empty vector by default. For backward compatibility we try calling
+    // export_state_interfaces() and only when empty vector is returned call
+    // on_export_state_interfaces()
+    return {};
+  }
+
+  /**
+   * Override this method to export custom StateInterfaces which are not defined in the URDF file.
+   * Those interfaces will be added to the unlisted_state_interfaces_ map.
+   *
+   * \return vector of descriptions to the unlisted StateInterfaces
+   */
+  virtual std::vector<hardware_interface::InterfaceDescription>
+  export_unlisted_state_interface_descriptions()
+  {
+    // return empty vector by default.
+    return {};
+  }
+
+  /**
+   * Default implementation for exporting the StateInterfaces. The StateInterfaces are created
+   * according to the InterfaceDescription. The memory accessed by the controllers and hardware is
+   * assigned here and resides in the sensor_interface.
+   *
+   * \return vector of shared pointers to the created and stored StateInterfaces
+   */
+  virtual std::vector<StateInterface::ConstSharedPtr> on_export_state_interfaces()
+  {
+    // import the unlisted interfaces
+    std::vector<hardware_interface::InterfaceDescription> unlisted_interface_descriptions =
+      export_unlisted_state_interface_descriptions();
+
+    std::vector<StateInterface::ConstSharedPtr> state_interfaces;
+    state_interfaces.reserve(
+      unlisted_interface_descriptions.size() + sensor_state_interfaces_.size());
+
+    // add InterfaceDescriptions and create StateInterfaces from the descriptions and add to maps.
+    for (const auto & description : unlisted_interface_descriptions)
+    {
+      auto name = description.get_name();
+      unlisted_state_interfaces_.insert(std::make_pair(name, description));
+      auto state_interface = std::make_shared<StateInterface>(description);
+      sensor_states_map_.insert(std::make_pair(name, state_interface));
+      unlisted_states_.push_back(state_interface);
+      state_interfaces.push_back(std::const_pointer_cast<const StateInterface>(state_interface));
+    }
+
+    for (const auto & [name, descr] : sensor_state_interfaces_)
+    {
+      // TODO(Manuel) check for duplicates otherwise only the first appearance of "name" is inserted
+      auto state_interface = std::make_shared<StateInterface>(descr);
+      sensor_states_map_.insert(std::make_pair(name, state_interface));
+      sensor_states_.push_back(state_interface);
+      state_interfaces.push_back(std::const_pointer_cast<const StateInterface>(state_interface));
+    }
+
+    return state_interfaces;
+  }
 
   /// Read the current state values from the actuator.
   /**
@@ -170,6 +257,16 @@ public:
     lifecycle_state_ = new_state;
   }
 
+  void set_state(const std::string & interface_name, const double & value)
+  {
+    sensor_states_map_.at(interface_name)->set_value(value);
+  }
+
+  double get_state(const std::string & interface_name) const
+  {
+    return sensor_states_map_.at(interface_name)->get_value();
+  }
+
   /// Get the logger of the SensorInterface.
   /**
    * \return logger of the SensorInterface.
@@ -190,11 +287,21 @@ public:
 
 protected:
   HardwareInfo info_;
+  // interface names to InterfaceDescription
+  std::unordered_map<std::string, InterfaceDescription> sensor_state_interfaces_;
+  std::unordered_map<std::string, InterfaceDescription> unlisted_state_interfaces_;
+
+  // Exported Command- and StateInterfaces in order they are listed in the hardware description.
+  std::vector<StateInterface::SharedPtr> sensor_states_;
+  std::vector<StateInterface::SharedPtr> unlisted_states_;
+
   rclcpp_lifecycle::State lifecycle_state_;
 
 private:
   rclcpp::node_interfaces::NodeClockInterface::SharedPtr clock_interface_;
   rclcpp::Logger sensor_logger_;
+  // interface names to Handle accessed through getters/setters
+  std::unordered_map<std::string, StateInterface::SharedPtr> sensor_states_map_;
 };
 
 }  // namespace hardware_interface